+++ daemon/execdriver/native/template/default_template.go
@@ -1,7 +1,7 @@
 package template
 
 import (
-	"github.com/dotcloud/docker/pkg/apparmor"
+	"github.com/dotcloud/docker/pkg/libcontainer/apparmor"
 	"github.com/dotcloud/docker/pkg/libcontainer"
 	"github.com/dotcloud/docker/pkg/libcontainer/cgroups"
 )
