+++ daemon/networkdriver/network_test.go
@@ -1,7 +1,7 @@
 package networkdriver
 
 import (
-	"github.com/dotcloud/docker/pkg/netlink"
+	"github.com/dotcloud/docker/pkg/libcontainer/netlink"
 	"net"
 	"testing"
 )
