+++ pkg/libcontainer/label/label_selinux.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/dotcloud/docker/pkg/selinux"
+	"github.com/dotcloud/docker/pkg/libcontainer/selinux"
 )
 
 func GenLabels(options string) (string, string, error) {
